backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
requests>=2.31
drf-yasg>=1.21
google-genai>=0.2.2
python-dotenv>=1.0.1
pypdf2>=3.0.0
